Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    float r,area;
+    double r,area;
     area = 3.14*(r*r);
-    printf("Area: %f",area);
+    printf("Area: %lf",area);
 
     return 0;
 }
